Playground: fix eject behaviour
